\name{RMoE-package}
\alias{RMoE-package}
\alias{RMoE}
\docType{package}
\title{
\packageTitle{RMoE}
}
\description{
\packageDescription{RMoE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RMoE}
\packageIndices{RMoE}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{RMoE}

Maintainer: \packageMaintainer{RMoE}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
